LPS-195477 Merger class MentionsMatcherRegistry with the class DefaultMentionsNotifier
